requirements.txt
slack-bolt>=1.21.0
slack-sdk>=3.33.0
pyyaml>=6.0
python-dotenv>=1.0
mcp>=1.0.0
watchfiles>=1.0.0
